remove watning.py file and move deprecated_warning into deprecation.py
